// QuestionComponent/QuestionnaireComponent/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    Question: ComponentFramework.PropertyTypes.StringProperty;
    Header: ComponentFramework.PropertyTypes.StringProperty;
    Col1: ComponentFramework.PropertyTypes.StringProperty;
    Col2: ComponentFramework.PropertyTypes.StringProperty;
    Col3: ComponentFramework.PropertyTypes.StringProperty;
    Col4: ComponentFramework.PropertyTypes.StringProperty;
    Col5: ComponentFramework.PropertyTypes.StringProperty;
    ShowInputField: ComponentFramework.PropertyTypes.StringProperty;
}
export interface IOutputs {
    Question?: string;
}